backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# PDF Processing
pdfplumber==0.10.3
PyMuPDF==1.23.8

# Excel Generation
openpyxl==3.1.2
xlsxwriter==3.1.9

# LLM Integration
google-generativeai>=0.8.0
# langchain==0.0.350  # Optional - not currently used
# langchain-google-genai==0.0.6  # Optional - not currently used

# Data Validation & Settings
pydantic==2.5.2
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# Database (Optional)
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# CORS
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
